requirements.txt
Appium-Python-Client==1.2.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-html==3.1.1
pytest-metadata==1.11.0
requests==2.26.0
selenium==3.141.0
toml==0.10.2
urllib3==1.26.6
